Patch 3 (#7)
* Update 03-eda.md (#530)

Added backticks for more readability and minor corrections, and improved the definition of the `np.log1p` function.

* Update 04-validation-framework.md (#531)

Added more information on some definitions and other minor changes to enhance clarity.

* Update 09-rmse.md (#532)

Minor changes to enhance clarity.

* Update 12-categorical-variables.md (#533)

Added a precision about variable appearing to be numerical when in reality, they are categorical.

* Update 16-summary.md (#534)

Added the notes taken for week 2

* Update README.md (#535)

Added the notes taken for Week 2

* homework 1 solution

* Update 13-regularization.md (#539)

* Added homework 3

* Update homework.md for 2024 (#538)

* Update README.md with Kemal's notes (#529)

* homework 4

* Q6 description fix

* foudn a minor error (#544)

* homework 2

* Update homework.md

* Update README.md with Kemal's notes (#549)

* typo in 14-explore-more.md (#548)

* Update formula (#550)

Formula was not rendering on GitHub and in VSCode

* Update 04-eda.md (#555)

fixed a small typo

* Update 03-validation.md (#554)

added details for the `train_test_split` function description with some corrections and fixed typos.

* Update 02-data-preparation.md (#553)

corrected some typos (grammar and code), added precision on some functions roles.

* Update 01-churn-project.md (#551)

Corrected some typos and enhanced precision.

* Update homework.md

* Update 02-accuracy.md

* Update 01b-saturn-cloud.md

* Update 03-confusion-table.md

* Update 06-auc.md

* Homework 4 - Questions 1: clarification (#564)

Adding more description to make it more clear.
Reject if you think this reveals too much or is incorrect.

* Small fixes

* Update 05-risk.md (#556)

Enhanced the clarity for the definitions of churn rate and risk ratio.

* Update 08-ohe.md (#557)

fixed typo in provided code and did small changes for easier understanding.

* Update 09-logistic-regression.md (#559)

Changed dead links into equations and fixed minor typos.

* Update 10-training-log-reg.md (#560)

Corrected some typos including code, and added a small precision.

* Update 12-using-log-reg.md (#561)

Fixed a typo

* Update README.md (#563)

Added the notes for this session

* homework 3 solution

* Added Notes and corrected small typos (#568)

Added the link to the notes taken as summary for this week and corrected two typos

* Update 06-auc.md (#567)

Added a small clarification

* Fixed dead links (#566)

Changed dead links to equations, and added small details for a definition

* Fixed a typo (#565)

corrected one typo

* Update for the confusion-table

I still think the table looks more accurate in my version

* Update 03-confusion-table.md, reverted html code

* Update 03-confusion-table.md

---------

Co-authored-by: Maximilien Eyengue <74904516+maxim-eyengue@users.noreply.github.com>
Co-authored-by: SVizor42 <timurkamaliev@gmail.com>
Co-authored-by: Vivien S. <73798748+viviensiu@users.noreply.github.com>
Co-authored-by: MathmoKiwi <96601542+MathmoKiwi@users.noreply.github.com>
Co-authored-by: Kemal Dahha <66003118+kemaldahha@users.noreply.github.com>
Co-authored-by: Alexey Grigorev <alexey.s.grigoriev@gmail.com>
Co-authored-by: Alexey Grigorev <alexeygrigorev@users.noreply.github.com>
Co-authored-by: Alex Kolodzig <alexkolo@users.noreply.github.com>